+++ frontend/packages/core/src/components/StackColumnChart3.tsx
@@ -201,7 +201,15 @@ const StackColumnChart = React.forwardRef<LineChartRef, any>(
                             },
                             axisLabel: {
                                 fontSize: 12,
-                                color: '#666666'
+                                color: '#666666',
+                                formatter:function (value:string, index:number) {
+                                    let str = ''
+                                    if (value.length > 6) {
+                                        str += value.substring(0,6) + '...'
+                                    }
+                                    
+                                    return str;
+                                }
                             }
                         }
                     ],
@@ -232,55 +240,7 @@ const StackColumnChart = React.forwardRef<LineChartRef, any>(
                             color: '#666666'
                         }
                     },
-                    dataZoom: [
-                        //给x轴设置滚动条
-                        {
-                            start: 0, //默认为0
-                            end: 20, //默认为100
-                            type: 'slider',
-                            show: true,
-                            xAxisIndex: [0],
-                            handleSize: 0, //滑动条的 左右2个滑动条的大小
-                            height: 8, //组件高度
-                            left: 50, //左边的距离
-                            right: 40, //右边的距离
-                            bottom: 0, //右边的距离
-                            handleColor: '#ddd', //h滑动图标的颜色
-                            handleStyle: {
-                                borderColor: '#cacaca',
-                                borderWidth: '1',
-                                shadowBlur: 2,
-                                background: '#ddd',
-                                shadowColor: '#ddd'
-                            },
-                            // fillerColor: new echarts.graphic.LinearGradient(1, 0, 0, 0, [
-                            //     {
-                            //         //给颜色设置渐变色 前面4个参数，给第一个设置1，第四个设置0 ，就是水平渐变
-                            //         //给第一个设置0，第四个设置1，就是垂直渐变
-                            //         offset: 0,
-                            //         color: '#1eb5e5'
-                            //     },
-                            //     {
-                            //         offset: 1,
-                            //         color: '#5ccbb1'
-                            //     }
-                            // ]),
-                            backgroundColor: '#ddd', //两边未选中的滑动条区域的颜色
-                            showDataShadow: false, //是否显示数据阴影 默认auto
-                            showDetail: false, //即拖拽时候是否显示详细数值信息 默认true
-                            handleIcon:
-                                'M-292,322.2c-3.2,0-6.4-0.6-9.3-1.9c-2.9-1.2-5.4-2.9-7.6-5.1s-3.9-4.8-5.1-7.6c-1.3-3-1.9-6.1-1.9-9.3c0-3.2,0.6-6.4,1.9-9.3c1.2-2.9,2.9-5.4,5.1-7.6s4.8-3.9,7.6-5.1c3-1.3,6.1-1.9,9.3-1.9c3.2,0,6.4,0.6,9.3,1.9c2.9,1.2,5.4,2.9,7.6,5.1s3.9,4.8,5.1,7.6c1.3,3,1.9,6.1,1.9,9.3c0,3.2-0.6,6.4-1.9,9.3c-1.2,2.9-2.9,5.4-5.1,7.6s-4.8,3.9-7.6,5.1C-285.6,321.5-288.8,322.2-292,322.2z',
-                            filterMode: 'filter'
-                        },
-                        //下面这个属性是里面拖到
-                        {
-                            type: 'inside',
-                            show: true,
-                            xAxisIndex: [0],
-                            start: 0, //默认为1
-                            end: 100 - 1500 / 31 //默认为100
-                        }
-                    ],
+                    
                     series: series
                 });
                 echart?.setOption(chartOptions, {notMerge: true});
